tests: Update parser tests to use the new names and simplify
